requirements.txt
djangorestframework==3.5.3
Django==1.10.5
psycopg2==2.6.2
markdown==2.6.7
django-filter==1.0.1
python-dotenv==0.6.2
djangorestframework-jwt==1.9.0
celery==4.0.2
django-celery-results==1.0.1
django-celery-beat==1.0.1
django-cors-headers==2.0.2